Create a new class for HomeScreen
